# api/openapi.yaml
openapi: 3.1.0
info:
  title: Library API
  description: HTTP API библиотечной системы (PostgreSQL + sqlx).
  version: 0.1.0
servers:
  - url: /
    description: current host
tags:
  - name: Health
  - name: Books
  - name: Readers
  - name: Loans
  - name: Search
  - name: Admin
paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/v1/books:
    get:
      tags: [Books]
      summary: List books
      operationId: listBooks
      parameters:
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Books page
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Books]
      summary: Create book
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookRequest"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          $ref: "#/components/responses/Error"

  /api/v1/books/{id}:
    get:
      tags: [Books]
      summary: Get book by id
      operationId: getBookById
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"

  /api/v1/readers:
    get:
      tags: [Readers]
      summary: List readers
      operationId: listReaders
      parameters:
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Readers page
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reader"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Readers]
      summary: Create reader
      operationId: createReader
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReaderRequest"
      responses:
        "201":
          description: Reader created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reader"
        "400":
          $ref: "#/components/responses/Error"

  /api/v1/readers/{id}:
    get:
      tags: [Readers]
      summary: Get reader by id
      operationId: getReaderById
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Reader
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reader"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"

  /api/v1/readers/{id}/loans:
    get:
      tags: [Loans]
      summary: List loans by reader
      operationId: listLoansByReader
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Reader loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
        "400":
          $ref: "#/components/responses/Error"

  /api/v1/loans:
    post:
      tags: [Loans]
      summary: Issue loan
      operationId: issueLoan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueLoanRequest"
      responses:
        "201":
          description: Loan issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"

  /api/v1/loans/{id}:
    get:
      tags: [Loans]
      summary: Get loan by id
      operationId: getLoanById
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"

  /api/v1/loans/{id}/return:
    post:
      tags: [Loans]
      summary: Return loan
      operationId: returnLoan
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Loan returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"

  /api/v1/search:
    get:
      tags: [Search]
      summary: Search books
      description: Полнотекстовый поиск книг через Elasticsearch.
      operationId: searchBooks
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: genres
          in: query
          schema:
            type: string
          description: Comma-separated genre names
      responses:
        "200":
          description: Search result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"

  /api/v1/admin/seed:
    post:
      tags: [Admin]
      summary: Seed database with fake data
      description: Требует заголовок X-Admin-Token (или Authorization Bearer), совпадающий с ADMIN_TOKEN.
      operationId: adminSeed
      security:
        - AdminToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedRequest"
            example:
              mode: reset
              authors: 100
              books: 1000
              readers: 500
              loans: 2000
              reservations: 200
              batch_size: 500
      responses:
        "200":
          description: Seed completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedResponse"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"

  /api/v1/admin/sync-search:
    post:
      tags: [Admin]
      summary: Reindex books into Elasticsearch
      operationId: adminSyncSearch
      security:
        - AdminToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                batch_size:
                  type: integer
                  default: 500
                recreate_index:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  indexed:
                    type: integer
                  duration:
                    type: string
        "401":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"

components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20

  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: not found

    Author:
      type: object
      properties:
        id:
          type: integer
          format: int64
        full_name:
          type: string
        birth_date:
          type: string
          format: date
          nullable: true
        bio:
          type: string
        created_at:
          type: string
          format: date-time

    Genre:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string

    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        isbn:
          type: string
        publication_year:
          type: integer
        pages:
          type: integer
        description:
          type: string
        created_at:
          type: string
          format: date-time
        authors:
          type: array
          items:
            $ref: "#/components/schemas/Author"
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"

    CreateBookRequest:
      type: object
      required: [title, isbn, publication_year, pages]
      properties:
        title:
          type: string
        isbn:
          type: string
        publication_year:
          type: integer
          minimum: 1000
          maximum: 2100
        pages:
          type: integer
          minimum: 1
        description:
          type: string
        author_ids:
          type: array
          items:
            type: integer
            format: int64
        genre_ids:
          type: array
          items:
            type: integer
            format: int64

    Reader:
      type: object
      properties:
        id:
          type: integer
          format: int64
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          type: string
          enum: [active, blocked, inactive]
        registered_at:
          type: string
          format: date-time

    CreateReaderRequest:
      type: object
      required: [first_name, last_name, email]
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string

    Loan:
      type: object
      properties:
        id:
          type: integer
          format: int64
        copy_id:
          type: integer
          format: int64
        reader_id:
          type: integer
          format: int64
        loaned_at:
          type: string
          format: date-time
        due_at:
          type: string
          format: date-time
        returned_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [active, returned, overdue]

    IssueLoanRequest:
      type: object
      required: [reader_id, copy_id]
      properties:
        reader_id:
          type: integer
          format: int64
        copy_id:
          type: integer
          format: int64

    BookDocument:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        isbn:
          type: string
        description:
          type: string
        publication_year:
          type: integer
        authors:
          type: array
          items:
            type: string
        genres:
          type: array
          items:
            type: string
        available_copies:
          type: integer
        indexed_at:
          type: string
          format: date-time

    SearchHit:
      type: object
      properties:
        book:
          $ref: "#/components/schemas/BookDocument"
        score:
          type: number
          format: double

    SearchResult:
      type: object
      properties:
        hits:
          type: array
          items:
            $ref: "#/components/schemas/SearchHit"
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer

    SeedRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [seed, reset, append]
          default: reset
        seed:
          type: integer
          format: int64
          default: 42
        batch_size:
          type: integer
          default: 1000
        authors:
          type: integer
        books:
          type: integer
        readers:
          type: integer
        loans:
          type: integer
        reservations:
          type: integer
        copies_min:
          type: integer
        copies_max:
          type: integer
        sync_search:
          type: boolean
          default: true
          description: After seed, reindex books into Elasticsearch

    SeedResponse:
      type: object
      properties:
        status:
          type: string
        mode:
          type: string
        authors:
          type: integer
        books:
          type: integer
        readers:
          type: integer
        loans:
          type: integer
        reservations:
          type: integer
        indexed:
          type: integer
        duration:
          type: string
        hostname:
          type: string